
\name{breastc}

\alias{breastc}

\docType{data}


\title{Gene Expression Data Example}


\description{
Microarray data set of van't Veer breast cancer.
}


\usage{data(breastc)}



\format{
A data matrix with 1213 genes and 97 samples.
}

\references{
Van't Veer, L.J., Dai, H., van de Vijver, M.J., He, Y.D., Hart, A.A. et al. (2002). Gene expression profiling predicts clinical outcome of breast cancer,\emph{Nature}, 415, 530-536.
}



\examples{
  
data(breastc)
  
head(breastc)

}


\keyword{datasets}
