\name{dlbcl}

\alias{dlbcl}

\docType{data}


\title{Gene Expression Data Example}


\description{
Log transformed Microarray data set of Rosenwald diffuse large-B-cell lymphoma.
}


\usage{data(dlbcl)}



\format{
A data matrix with 661 genes and 141 samples.
}

\references{
Rosenwald, A., Wright, G., Chan, W.C., Connors, J.M., Campo, E. et al. (2002). The use of molecular profiling to predict survival after chemotherapy for diffuse large-B-cell lymphoma, \emph{New Engl. J. Med.}, 346. 1937-1947.}



\examples{
  
data(dlbcl)
  
head(dlbcl)

}

\keyword{datasets}
