\name{exploreCalc}
\alias{exploreCalc}
\title{ The exploreCalc function
}
\description{calculates the 'mean', 'median','variance', 'MAD', and 'quantile' for biclusterd and. clusterd data.}
\usage{
  exploreCalc(bic.mat)}
\arguments{
  \item{bic.mat}{Biclusterd or clustered data matrix.}
}
\details{
  an internal function used to calculate the summary statistics of biclustered and clustered data. it supports for \code{\link{exploreBic}} function. It is used only for internal purpose.
}
\value{
  Summary statistics('mean', 'median','variance', 'MAD', and 'quantile') for biclusterd and clusterd data.
}

\author{Mengsteab Aregay \email{mycs.zab@gmail.com}}

\seealso{
  \code{\link{exploreBic}}}

\keyword{exploreCalc}
